resolve compile error
The following error is generated when compiling with proton-clang:

WARNING: EXPORT symbol "gsi_write_channel_scratch" [vmlinux] version
generation failed, symbol will not be versioned.

The cause is that genksyms does not understand the return type
__packed of the function __gsi_update_mhi_channel_scratch. Since the
return type is not necessary, we can just drop it from the function
and resolve the error.

https://stackoverflow.com/a/66250712